+++ Cargo.toml
@@ -5,13 +5,14 @@ rust-version = "1.64.0"
 edition = "2021"
 
 [features]
-default = ["debug", "cli"]
 cli = ["comfy-table", "clap", "enable-serde", "metrics-tracing-context", "metrics-util", "tracing-subscriber", "web-app"]
 debug = []
+default = ["debug", "cli"]
 enable-serde = ["serde", "serde_json"]
-web-app = ["tokio", "tokio-stream", "tokio-util", "axum", "axum-server", "hyper", "hyper-tls", "tower", "tower-http"]
 self-signed-certs = ["hyper-tls"]
+shuttle = ["test-fixture"]
 test-fixture = []
+web-app = ["tokio", "tokio-stream", "tokio-util", "axum", "axum-server", "hyper", "hyper-tls", "tower", "tower-http"]
 
 [dependencies]
 aes = "0.8"
@@ -58,6 +59,7 @@ x25519-dalek = "2.0.0-pre.1"
 criterion = { version = "0.4.0", default-features = false, features = ["async_tokio", "plotters", "html_reports"] }
 iai = "0.1.1"
 proptest = "1.0.0"
+shuttle = "0.4.1"
 
 [lib]
 name = "raw_ipa"
