+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "shuttle")]
+mod randomized;
